parser and lexer work together? maybe
so changed the parser slightly to have a STALL state where
it will chanage states but not reduce or take next input.
a shift with no input grabbing.

now code will run in the middle of productions or at the end
of productions.

still some other changes i cant remember at this time.